requirements.txt
beautifulsoup4==4.10.0
Flask==2.0.1
line_bot_sdk==1.20.0
numpy==1.21.1
pandas==1.3.0
pymongo[srv]
pymongo
requests==2.26.0
scikit_learn==1.0.2
keras
gunicorn
tensorflow-cpu
